Fixed glitch that had floats for (x,y,z) entity position rather than doubles
git-svn-id: svn://svn.code.sf.net/p/open-dis/code@405 1da32bc5-2e4a-4c7f-92e1-9780f3f753aa

+++ trunk/open-dis/cpp/DIS/FastEntityStatePdu.h
@@ -84,13 +84,13 @@ protected:
   float _zVelocity; 
 
   /** X value */
-  float _xLocation; 
+  double _xLocation; 
 
   /** y Value */
-  float _yLocation; 
+  double _yLocation; 
 
   /** Z value */
-  float _zLocation; 
+  double _zLocation; 
 
   float _psi; 
 
@@ -207,14 +207,14 @@ protected:
     float getZVelocity() const; 
     void setZVelocity(float pX); 
 
-    float getXLocation() const; 
-    void setXLocation(float pX); 
+    double getXLocation() const; 
+    void setXLocation(double pX); 
 
-    float getYLocation() const; 
-    void setYLocation(float pX); 
+    double getYLocation() const; 
+    void setYLocation(double pX); 
 
-    float getZLocation() const; 
-    void setZLocation(float pX); 
+    double getZLocation() const; 
+    void setZLocation(double pX); 
 
     float getPsi() const; 
     void setPsi(float pX);
